Add license files mandated by (L)GPL.
Currently only qtbase includes the required LICENSE files, which
are referred to from the code headers. The same files should
be distributed as part of all the modules as well, so that the
(L)GPL licenses are adhered.

These files were wget:d from:

https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LGPL_EXCEPTION.txt
https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LICENSE.LGPL
https://qt.gitorious.org/qt/qtbase/blobs/raw/stable/LICENSE.GPL

(in qttranslations, no FDL licensed files so LICENSE.FDL not needed)

Change-Id: I5d50fe74e735b0f6395f0fcd53f06c7022e066ca
Reviewed-by: Jason McDonald <macadder1@gmail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>