revert addition of declarative and quickcontrols to meta catalogs
the meta catalogs are there for qt4 compat. consequently, there must be no
quickcontrols. declarative was also excluded, because is was not in qt4,
either - only quick1 was, which is already part of the meta catalog
(the split-qt-ts.pl script tries to recycle translations from quick1 for
declarative, but that has no relevance for the meta catalog).

Change-Id: I0b33f7cd6cbd51d478d53849bd2ce8657df34539
Reviewed-by: Lars Knoll <lars.knoll@digia.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>