Fix regression from Qt4 with layout direction
In order to resolve Right-to-Left layout from the translation file, we
used to check if the result of QApplication::tr("QT_LAYOUT_DIRECTION")
was the string "RTL".

During modularization this changed context (using the wrong
QCoreApplication::tr()), but the translation files didn't change.

It turns out that it didn't make much sense to use that context,
since the layout direction can only be changed on the QGuiApplication.

We therefore have changed the context to QGuiApplication (in qtbase),
and we must therefore update the relevant translation files.

Change-Id: Iacea23da32a3d4128c259d7e8aad1823801dfd5f
Task-number: QTBUG-43447
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>