Port of QtTranslations to CMake
Initial port of QtTranslations to CMake.

As we discussed, since we can't read the CMakeLists.txt files to extract
information about the source files and include directories we will
settle for globing all the source files and feeding that as an input
to lupdate.

This port also forgoes the sh/bat commit-ts script in favor of a CMake
script that can be executed on both platforms. This comes at the expense
of having to call git commit manually since it's not possible to
interact with the git prompt when executing the script.

This patch does not cover exclusions. These will be addressed
as part of future work.

Change-Id: I4c1d5f6209b42405b2c2a05f25fc5aa645122990
Reviewed-by: Qt CMake Build Bot
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>