Update commercial license headers
Updated header.COMM to the files with GPL-EXCEPT license header
in tqtc-qttranslations.

The commercial license header may contain some additional
lines so that its line count equals with the earlier
license header. Reason for this is that some autotests use
hard coded line numbers and a change in the line count causes
failures in tests.

Task-number: QTQAINFRA-4212
Change-Id: I6a71caed50a34b0d6b38f6b5cfdb625df497050e
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>