Update commercial license headers
Updated header.COMM to the files in tqtc-qttranslations.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4956
Change-Id: I44694ff84addafd19ba43888c08e53127212c8b6
Reviewed-by: Antti Kokko <antti.kokko@qt.io>