CMake: Ensure directory creation for Unix Makefiles generator
In contrast to the ninja generator, the unix makefile generator does
not ensure a directory is created before a byproduct file (.qm file)
of a custom target (updateqm) is created.

This resulted in a possible race condition where sometimes the
directory was created by a different rule, and sometimes not, leading
to a build failure like

 translations/qtwebsockets_en.qm': No such file or directory

Create a directory explicitly as part of the custom target.

This is an issue only in 6.2, because
c17fc36d54b247e5aee9e609ac91687a9548c257 was never picked to 6.2
(despite the commit footer).

Amends 8c585f7b0f291463934cc4ffbddd925eff709f13

Fixes: QTBUG-109973
Change-Id: I564561f98cc385a2fcfd4f053583795658edabe5
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>