Remove check-ts target
The target is broken after the tranlsations.pro removal. It's not clear
what is the initial purpose of the script. The logic in perl scripts
can be easly ported, but since it's not in use for a long time and
no one noticed removal looks more correct.

Also remove the scrips related to this target.

Pick-to: 6.5 6.2
Change-Id: Iff9db0bf0e6dbe0abbbae0fb3c55ef844b6d9059
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>