Exclude 3rdparty files from the list of catalog files
file(APPEND is significantly slower in CMake 3.27.1, most likely due to
a regression in CMake. As a result, appending the result of globing to
file_list can be very slow, e.g., when processing qtwebengine. With a
bit of rework, we can eliminate the need for APPEND-ing, and later
exclude the /3rdparty/ files since we don't need them anyway, similar to
doc/snippets/.

Pick-to: 6.5 6.6
Fixes: QTBUG-115634
Change-Id: I34fa9dcd68c488c05a68aa69835dc59885f8599d
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>